digest: fix: leading zeroes were truncated from sha384 string representation (#208)
@StephenDiligent figured out that creating a sha384 digest and converting it's
string representation back to a digest can fail with the error:
"hash length is 95, expected length 96".

If a calculated digest started with 0, the leading 0 was truncated from the
hash. This happened because the hash was stored as big.Int, leading zeroes can
be omitted for an integer without changing the meaning. Truncating a zero from a
digest makes it invalid.

Fix it by using the functions from the hex package to create a string
representation of the hash instead of big.Int.

A testcase is added for the bug.
The existing testcase are changed to compare digests via their string
representation instead of converting the Sum to a string on their own.