gitfileinputs: error out when one of the paths did not match any files
Instead of passing the parameter "--error-unmatch" to "git ls-files",
"error-unmatch" was passed as pattern.

This caused that when a GitFileInput Paths list contained multiple elements, it
was sufficient that only one of the patterns matches files to not cause an
error.
When specified path does not match baur should fail with an error instead.

This fix is a backport from master.