tests: redirect command output to testlogger
Ensure that all testcases call redirectOutputToLogger(t) (or initTest()) to
ensure all output is logged via the test loggers.

Some testcases were calling initTest() not in the subtest but in the parentTest.
initTest() must be called in the subtests, if a subtest fails FailNow() is
called. When the function is called on the t.Testing object from the parent, the
subsequent subtests will also not run.