tests: introduce io.Writer test logger wrapper
Outputs of command that were run as part of a test could not be associated with
the executed testcase from go test and were shown intermingled with output from
other testcases.

Introduce an io.Writer that logs all writes to t.LogF() and writes them to
another chained writer.
The writer is used for the stdout and stderr streams of the commands to enable
logging the output via the test loggers.
This ensures that go test can associate the output of commands with the executed
testcases.