tests: support validating exit codes when using initTest()
To be able to also use the initTest() function when the exit code of a command
needs to be validated it is refactored to set the exitFunc() to a function that
panics instead of calling t.Fatalf().

To panic() a struct is passed that contains the Exitcode with that exitFunc()
was called.

This allows to recover from the panic and validate if the exitCode is as
expected.