log: remove Fatal functions
remove the Fatalln and Fatalf functions from the logger.

Fatal errors should always result in a message to the users which should be
printed via the stderr stream.
The Fatal log functions were also not terminating the program via the exitFunc()
which caused different behaviors when they were run in testcases.