loader: fix paths containing dots not correctly interpreted
If an app or tasks were loaded by specifying a path and the path contained dots,
e.g. "../", the loader failed with the error that the specifier contained too
many dots.

This is fixed by first checking if a specifier is the path of an existing
directory that contains an .app.toml file, before interpreting it as app-name.

The loader was refactored to share more code between LoadTasks() and LoadApps().
Parsing the  specifiers is moved to an own function (parseSpecs()).