docker: improve choosing credentials for registry, support URLs
"docker login" supports to specify registry per hostname and as URL.
If the credentials used the URL, baur was not able to find the credentials when
the push was done using only the hostname of the registry server.
This is a common usecase for GCR auth per json_key.

The authentication logic was also approved in the following ways:
- registry URLs with ports are supported,
  If a registry was authenticated with the default port 5000), and no port is
  given for the registry url in the app config, the credentials for the
  registry with the default port is used
- the credentials for the default registry (index.docker.io) are only used when
  the host of the specified registry matches or no registry is given in the app
  config
- when no credentials were found for registry, none are used instead of the
  first found one, this was a security issue, credentials might have been send
  to a wrong server