exec: support to specify environment variables for command execution
Support to specify environment variables for command execution.
It works analogous to how it works the stdlib exec package.
Environment variables are specified as []string, if the slice is nil the
variables of the current environment are used.

This will be required to run "go env" for the gosource resolved with userdefined
env vars.