gosourceresolver: use $GOMODCACHE in paths for non-vendored external sources
When a go.mod project did not vendored external libraries, baur resolved the
source files of the libraries to the user's $GOMODCACHE directory.

The $GOMODCACHE is normally in the home directory of the current user and
outside of the baur repository.

This means when the same application was build with another user or a different
GOMODCACHE setting, the files resolved to different paths, baur rebuilded the
applications unnecessarily.

To make those paths independent of local paths, the gosourceresolver
replaces paths that have start with the value of the $GOMODCACHE variable with
"$GOMODCACHE".