cfg: support to use environment variable values in fields
The values of environment variables can now be used in most fields of
the app config files.
Referencing environment variables works via the default template syntax instead
of the existing variable syntnax.

The other supported variables will be changed to also use the template syntax
instead.
The variable syntax does not work here because it lacks a character that
indicates the end of the string.

"{{ env <VARNAME> }}" in an app config field is replaced the value of the
environment variable VARNAME.
If the environment variable is not defined an error will occur.