cfg: do not document supported variables in example app config
In the generated example app config file the inline documentation were
mentioning the supported variables of the field.

The information was redundant and it was hard to maintain, when
support for a new variable was added many comments had to be kept uptodate.

The list of supported variables can be instead either documented in a
header comment in the file or in the wiki.