resolver: use same glob syntax for Glob and GitPath resolver
The Glob and GitPath resolvers were processing patterns differently.
For example when a '*' was used as path in the Input.GitFiles section it also matched
directory separators, when '*' was used as path in the Input.Files section it
did not match directory separators.

Ensure paths are resolved the same by resolving globs for both resolvers via fs.Glob().
The GitPath resolver is then passing the paths to "git ls-files" to check if
they are committed to the repository. The "--noglob-pathspecs" parameter is
passed to "git ls-files" to treat all paths literally.