gosourceresolver: whitelist LocalAppData, USERPROFILE windows env variable
On Windows the LocalAppData environment variable is used to determine the
default GOCACHE directory.
If it's not set and neither GOCACHE is, "go env" fails with:
"missing $GOPATH" or "build cache is required, but could not be located: GOCACHE
 is not defined and %LocalAppData% is not defined" error

USERPROFILE is used to determine the default GOPATH if it's not set and neither
GOPATH is, "go env" fails with: "missing $GOPATH".