resolver: gosource: whitelist LocalAppData windows env variable
On Windows the LocalAppData environment variable is used to evaluate
default go directories like GOCACHE and GOPATH.
Set the LocalAppData when running "go list" for the gosource GitFiles resolver.
This ensures that go can calculate the default directories correctly.

This should fix resolving Gitfiles failed on Windows when GOCACHE or
GOPATH were not set explicitly.