gitpathresolver: fix: error was returned when not all files were git tracked
When a glob like "*" was passed to the gitpathresolver and not all of the
filepaths to that the glob resolved were tracked by git, an error was returned.

This changed accidentally the previous behavior and is not wanted.
The resolver should return only files that are in the filesystem **and** tracked
by Git.

This is fixed by running the resolver per entry in the Inputs.GitFiles.Path
field in the app config and not passing "--error-unmatch" to git-ls files.
The interface of the resolver is made more similar to the FilesResolver.
It now accepts only a single path instead of a list.
Checking if an error returned because a glob resolved to 0 files is also done in
the caller instead of in the resolver.