gosourceresolver: do not return files that are in the GOCACHE directory
When a go test already run before, the gosourceresolver was returning paths
from the local GOCACHE directory.

Do not return files that are in the GOCACHE directory. We only want to return
source code files. Files in GOCACHE depend on the source files and are
regenerated reproducible for the same source files. They do not need to be
tracked as inputs.