git: fix: discovering input files fails with: argument list too long
If a lot of files were resolved it could happen that running "git ls-files"
failed with: "/sbin/git: argument list too long".

When the length of the strings of the pathspecs slice that is passed to
git.LsFiles() exceeds a threshold, the function is now splitting the number of
arguments in smaller slices and invokes "git ls-files" for each of it.

As threshold "32767 - 2048" is chosen.
Windows has a limit of 32767
(https://devblogs.microsoft.com/oldnewthing/20031210-00/?p=41553).

In the exec sycalls the array for the command line arguments and the environment
variables share the same limit. Therefore 2048 is subtracted to reserve some
space for environment variables (as recommended in
https://www.in-ulm.de/~mascheck/various/argmax/).
Modern Linux/Unix systems have a much higher limit for the argument size.

The issue can still happen depending on the OS, e.g. on older Linux systems or
Linux systems with a small stack size.
It's unlikely that baur is used on those system.
The only safe way to set the limit would be to evaluate the max. length during
runtime by running a command and extending the number of parameters until the
error happens. This seems to be over the top for the current usecases of baur.