cfg: do not use different set of variable resolvers for include fields
For resolving variables in config files 2 different sets of resolvers were used:
- for the "Includes" field in config only resolvers for ' {{ .app }, {{ .root }}
  and {{ .env <VARNAME> }}
- all available for all other field

Remove the exception for the "Includes" field, use the same set of resolvers
then for all other fields. This simplifies the  code a bit and does not have any
real disadvantages.

This means {{ .gitcommit }} and  {{ .uuid }} variables can now be used in the
"Include" field, there is no valid usecase for it though.