flags: support to specify a default value for flag fields
The Fields flag now supports to specify it's initial default value.
So far the default value was always the same then the supported fields.
This allows to support fields for the "baur status" and "baur ls apps"
command that are not shown by default.

In the usage output the supported fields are now also always printed in the same
order.
Previously the order differed from invocation to invocation.