inputs: fix: input digests differ when baur repo is in a different path
The digest that was calculated for input files was adding the absolute file path
to the data for that the digest is calculated.
This caused that when the baur repository was in a different path, the input
digest of the files differed.

Use the repository relative path to the file for calculating the digest.
The File struct is removed, InputFile and OutputFile are not based on it
anymore. The used shared functionality became too less to share.