diff inputs: add --input-string inputs only to first specified task/run
When the "input-string" parameter was passed to "baur diff inputs", the input
string was added to both targets that were specified.
Both runs always had the additional input string and therefore could not differ
regarding the string input.

Change the behavior to addd the string input that is specified as parameter only
to the first target specified by the arguments.