storage: refactor input string representation
File Input and String Inputs were represented by the same storage struct
and also stored in the same postgresql table.
In the input.name db column and in the storage.Input.Name struct field was
stored for file inputs the path, for string inputs the string with a "string:"
prefix.
The input table had a unique constraint for the name and digest column, that was
used to to a database upsert operation (only insert if records does not exist).
When long string inputs were specified as parameters for "baur run" it happened
that the input could not be stored in the table because it was too long to be
stored in the unique index (index row requires 10032 bytes, maximum size is 8191
(SQLSTATE 54000)).

To fix this issue and to have clear struct field and db column names the storage
representation is refactored.
Input Files and Input strings are now represented by separate structs and stored
in separate tables.
The unique index for the input_string table is only created for the digest field
and not the string itself anymore. The digest depends only on the string.