tests: reset log, stdout, stderr streams when a test finishes
When a testcase was calling redirectOutputToLogger() and the following executed
testcase wasn't, all it's output was redirected to the logger for the first
testcase.

These log messages could be lost or associated with the wrong testcase.

Fix it by resetting the log streams that were changed to their previous values
when the test finished execution.