loader: fix: a task spec with only a wildcard for the app loaded all tasks
If a task specification was passed to "baur status" or "baur run" that consisted
of a wildcard for the app and a specific task-name like '*.build', all tasks for
all apps were loaded,  instead of only all tasks matching the task-name.

Fix it by removing the specs.allApps fields.
It is not necessary because if a spec is parsed with an non-empty task part an
element for it is added to specs.taskSpecs and it will then be loaded by
Loader.tasks().