tests: check in TestRunSimultaneously also if parallel task processes exist
Extend the TestRunSimultaneously testcase to also check if the tasks are really
executed in parallel.

The parallel executed tasks run a shell script that writes their start and exist
timestamp to a logfile.
After "baur run" was executed, the tests checks if all timestamps of the
executed tasks are in the same interval.