run: check if tasks are run in parallel via timestamp in testcases
Neither ps nor tasklist is able to show on Windows the whole cmdline of running
tasks.
This makes it impossible to use it to figure out if all processes are running in
parallel, all parallel running tasks are shown as "bash".

Instead log the start and endtimestamp of the tasks to files and check if they
are all in the same interval.
This should work on Windows and Linux.