routines: fix: race on pool shutdown
There was a chance that when Wait() and Queue() was called in parallel, that
the work was not executed before the pool terminated and neither Queue() paniced
because it was called after the pool terminated.

Hold the lock while dequeuing work from the queue and checking if p.terminate is
set.