run: rename func outputsExist(9 to declaredOutputsExist()
The previous function name sounded like it would return false if no ouputs were
defined and no outputs existed after the task run.
This is not the case.

The new name should it make it more clear that the function returns false if the
outputs do not exist as declared in the configuration of the task.