vendor: update aws-sdk-go to aws-sdk-go-v2
Upgrade the awsk-sdk-go to version 2.
The same client configuration is kept.
Messages from the aws-sdk-go-v2 are logged now with a "aws-sdk-go-v2" prefix.

The s3.Logger.DebugEnabled() method of the logger interface was removed, to
simplify the code. The s3 client logs so less that those few unnecessary logging
calls do not matter.

The client.Upload() method is not changed to pass a context to the S3 uploader
because it would require to make a downward incompatible change in the public baur
API interfaces and methods of pkg/baur/uploader.go
Making use of the context is postponed to a future major release.