ci: use github actions instead of circleci
Running the tests on Windows CircleCI machines was very slowly. It took ~15min,
after switching to the newest windows orb + installing mingw additionally.
The main reason was the installation process on the windows machines.
The github action windows image already have everything installed what is needed
(psql, gcc) and are available almost instantly.

Switch to using github actions for CI.
The CI jobs now also run the tests with CGO disabled.
When CGO is disabled using the race detector is not possible. The release are
build with CGO disabled, therefore we also run the test without cgo now.

In CI currently only the Linux baur binary is build. This should be sufficient.
I never had the case that I could compile baur on linux but not for another
architecture or OS.

The circleci bade from the readme is removed.