repository: use canonicalized repo and cfg path
When the baur repository was in a path containing a symlink, path to input files
were not relative to the realpath of the repository directory.
For example:

/tmp/baur-repo/inputfile
/tmp/baur-repo-symlink -> /tmp/baur-repo (symlink)

If baur was invoked in /tmp/baur-repo-symlink, input files path were like:
"../../tmp/baur-repo/inputfile, instead of "./inputfile".

Depending on if baur was invoked in the realpath or the path containing a
symlink, the input file paths differ and could cause unnecessary rebuilds.

This is fixed by evaluating all symlinks and using the absolute path of the
configuration file path passed to baur.NewRepository()