fix TestDifferencesOutputWithCorrectState failed on Windows
In the Windows CI machine os.MkdirTemp() returned a short DOS Path to the temp
directory used as repository directory.
When input files were resolved, the long path to the repository directory was
used.
This caused that the paths used different representations of the same directory
in TestDifferencesOutputWithCorrectState(), comparing their string
representation failed, they were not equal.

This is the same issue that was fixed in commit
"repository: use canonicalized repo and cfg path" for the Repository struct.