git: add functions to get untracked files in repository
- Add a function UntrackedFiles() that runs "git status" and returns the list of
  untracked files. In the "git status" output, lines about untracked files start
  with "?? ".
- Add a function repository.WithoutUntracked() that takes a list of paths in a
  repository, and returns a new list with all untracked files removed.
  The function uses UntrackedFiles().
  As other functions of the Repository struct, it caches the return value of
  "git status" and only runs it once.

This functionality will replace using LsFiles to figure out if files are tracked
or untracked in a repository.
It is much faster because "git status" it is only invoked once and then caches
the list of untracked files.

LsFiles worked the reverse way. It had to be called multiple times per list of
files to figure out which of those are tracked.