inputs: improve performance of resolving of git_tracked_only file inputs
Use the new git.WithoutUntracked() method to figure out if files are tracked or
untracked in a git repository.
This speeds up resolving of inputs that have the git_tracked_only flag set.
Enabling git_tracked_only does not have a noticeable performance impact anymore.
It only runs "git status" 1x per baur command invocation and then checks if the
resolved file inputs are in map of untracked files.
Before "git ls-files" was called at least once per unique Input file
specification that had git_tracked_only set.

To achieve that the VcsState object is now passed to the InputResolver on
creation. This is a breaking change, it changes the public baur package API.

The gitpath resolver was removed, it's functionality was integrated into the
inputresolver. Since baur 1.0, git_tracked_only became a flag for the InputFiles
definition.
The GitPath resolver was calling the glob resolver and then filtering out all
untracked files. Now the git_tracked_only flag is handled similar like the Optional flag, in the
InputResolver.