inputresolver: cache glob file results per path
An FileInput section in a config file, can contain multile Path specifications.
The input resolver cache was caching results for a while FileInput section.
Each path in a FileInput section was resolved separately though.

Multiple apps can have FileInputs sections that contain the same path with the
same options.
To improve caching, cache the results per path of a FileInput.