inputfile: use singleton pattern for input file
Instantiate an InputFile struct only once per file.
Previously if multiple apps had the same file defined as InputFile, for each
app an InputFile was created that calculated the digest.

To prevent calculating the digest multiple times for the same InputFile,
instantiate InputFiles only once per path.

This is achieved by introducing an InputFileSingletonCache struct, that has a
map of absolutePath->*Inputfile. InputFileSingletonCache. CreateOrGetInputFile
returns an *InputFile from the cache if one was created for the same path
before, otherwise it creates one and adds it to the cache.