inputfile: accept absolute path instead of repoRelPath in constructor
Change NewInputFile() InputFileSingletonCache.CreateOrGetInputFile() to accept
the absolute path to the file instead of the absolute repository path and relative file path.
The methods created the absolute path from both of these arguments and only used
the absolute and relative path.

The caller (InputResolver) already has the absolute path and can pass it
directly. This prevents redundant filepath.AbsPath() calls.

This is a breaking change. It changes the baur package API.