inputfiles: do not resolve directory symlinks
When an InputFile Glob matched a directory that was a symlink, the symlink was
resolved and the resolved directory path was returned.

Do not resolve directory symlinks. This can hide changes in the InputFiles of an
Application and prevent that a task is rerun.

If a task has as Inputfile the real file path defined and a path which contains
a glob + a directory symlink, only the real file path was returned in the
result.
If the input file with the directory symlink was removed or added, the inputs
did not change and would not have triggered a task rerun.

This fixes the TestResolveSymlink testcase failure since d2f8b31.