inputs: support tracking environment variables
This commits adds support to track the value of environment variables as Input.
As all other inputs, they can be defined in an include or app configuration
file.
The .app.toml configuration sections looks like:
  [[Task.Input.EnvironmentVariables]]
    names = [
      "APP_VERSION",
      "PAYMENT_APP_*",
    ]
    optional = false

Environment variables names can contain go glob patterns. This allows to track
all variables that a certain patterns without having to list each of them.
If the optional flag is true, baur command that calculate app inputs (baur
status, baur run, etc) will fail if no environment variable is defined matching
the defined pattern.
Declared but undefined environment variables are treated as non-existing.

The values of environment variables is not stored in the database or shown by
"ls inputs", to prevent that secrets are accidentally shown or stored.
A new database migration is added that adds table to store the digests of
environment variables of runs.