storage: support applying schema migrations
This commit adds support to change the database schema via migrations.

Schema migrations are stored in the embedded FS pkg/storage/postgres/migrations.
They are named as <VERSION>.sql. <VERSION> must be an int32 >=1.
The initial db schema that was stored as a constant in schema.go is moved to
pkg/storage/postgres/migrations/1.sql.

The migrations files are read and the filename is parsed when Init() or
ApplyMigrations() is called.
Init() applies all found migrations files.
ApplyMigrations() only applies the migrations that have a version that is
smaller then schema version stored in the database.

ApplyMigrations() will be used in a following commit.