commands: implement baur upgrade db
Provide the new command "baur upgrade db", to upgrade the database schema to a
newer version.

If the database schema is already uptodate, the command succeeds with a
descriptive message.
If the database does not exist, the command fails with a message saying that
"baur init db" needs to be run to create the database first.