inputs: support to exclude files by glob patterns
Support to exclude files from the inputs by specifying them in an ignore list.
The list of files to exclude can be specified as glob patterns in an
"Input.ExcludedFiles" section. Patterns in the ExcludedFiles section that
reference non-existing paths are silently ignored.
The ExcludedFiles are processed after all other input sections.
Input files that were discovered and match one of the ExcludeFiles patterns are
removed.
ExcludedFiles sections can be defined in include files and imported on task
level.

In the .toml config file the Files, GolangSources and EnvironmentVariables
members of the input sections are lists.
E.g. '[[Task.Input.Files]]' sections for a task can exit with different options.
The ExcludedFiles is a single element, because it does not have additional
options.
Because it is not a list, the representation in the toml config differs:
'[Task.Input.ExcludeFiles]' vs '[[Task.Input.Files]]'

Example .app.toml section:
  [Task.Input.ExcludedFiles]
    paths = ["dbmigrations/1.sql"]