completion: complete "ls apps" arguments
Add a custom shell completion function for "ls apps" arguments.
When the current directory is part of a baur repository, the argument is
completed to app-names and relative app directories.
The completed app directories are relative to the current working directory.

The following arguments are suggested by bash when "baur ls apps <tab>" is
entered in the root of the baur-examples git:
    a-team/go/src/github.com/simplesurance/hello-server  project-pink-einstein/rng
    a-team/myredis                                       project-pink-einstein/unixtime
    hello-server                                         rng
    myredis                                              unixtime