init bashcomp: do not use $HOME/.local/share on windows
Do not fallback on windows to use $HOME/local/share as XDG_DATA_HOME.
The directory is unix-specific and will not work Windows, maybe with the
exception when using cygwin(?).

To be safe, require either BASH_COMPLETION_USER_DIR or XDG_DATA_HOME to be set
to autodetect the completion directory.