completions: support completing "baur show" arguments
Add support for completing the argument of "baur show".

Run-IDs are currently not completed.
It would be possible, by querying the existing IDs from the database that start
with the same numbers then the ones on the commandline. If a lot of records
exist, doing e.g. "baur show 1<TAB>" results in thousands of options.
That does not seem very helpful nor necessary.