release: switch to using goreleaser
Configure goreleaser to create archives and upload them to github, it's smoother
then the previous process.
The archive names and content are the same then before.
The tar archives are not reproducible anymore, the owner of the files is not
0/0, the timestamp is not fixed. I could not find a way to do it in goreleaser.
It is only possible to set owner and timestamp for additional files, not the
main binary.
The archive for windows is now in .zip format, the format is more common on
Windows.
Releases are now also created for windows/arm64.
The Makefile targets that were used before to create releases were removed.