goreleaser: validate that git-tag and version in ver are the same
In the 3.0.0-rc1 release binary the string in the ver file did not match the
git-tag.
I forgot to set the string in the ver file.

Add a before hook to goreleaser, that checks if a git-tag exists and matches the
string in ver file. If it does not, the script fails and goreleaser aborts with
an error.
This will prevent that it happens again.