test: prevent deadlock termination in forever blocking child process
When a go program hangs forever, the go runtime detects it and terminates it
with the error message:
	fatal error: all goroutines are asleep - deadlock!

This happened in the TestProcessTerminatesWithParent, where a child process was
only running `select {}` to wait forever until it gets killed.

Replace select{} with a very long time.Sleep() call in the
TestProcessTerminatesWithParent() to prevent that the testcase fails because the
child process gets terminated because of the described reason.