exec: only stream output when log function is set
By default the log function was an function that did nothing.
If logFn was not nil, the output was always splitted in lines and logged via
logFn. This is unnecessary.

Instead of setting DefaultLogFn to an noop function, set it to nil.
This allows to determine if logFn is set and prevent unnecessary processing of
the stdout and stderr of the process.