exec: fix: execution fails when output is streamed and lines are >=512KB
If the output of a command was streamed via the logFn function and the output
did not contain newlines after latest each 512Kb, the execution failed.
This happened because when splitting the output in lines failed, the PipeReader
was closed, causing all writes to the MultiWriter to fail. Which in turn causes
writes from the process to stdout and stderr fail.

Handle this error case gracefully. Streaming output is for debugging purposes
and not required for the result.
If the line scanner returns an error, log it and discard all further output of
the logreader.
If discarding the output fails, the logReader is closed with an error. This
should not happen, if it does and the logreader would not be closed, it would
cause the process to block forever.

A testcase for the error scenario is added.