exec: fix: command output not shown when MultiWriter fails
When an error happened writing the output of the command to the MultiWriter, the
last output is missing in the error message.

Reorder the MultiWriters, to first write to the stdoutPss, stderrPss writer that
are the buffers for the error message.
That way the buffers contain the last output if a write operation to the
following writers fail.