exec: fix: duplicate newlines in streaming output
When "--verbose" was passed to "baur run" the logger was used as logf function.
The logger appends a "\n" when the last character is not already a newline.
Because "\n" character was colorized with the logFnStderrStreamColorfn, the
logger did not detect the newline anymore and appended a second one.

Do not pass "\n" to c.logFnStderrStreamColorfn.