s3: Store s3:// instead of https:// in db for uploaded objects
For artifacts that are uploaded to S3 baur now stores it s3://BUCKET/OBJECT-ID
URI in the database, instead of the path-style S3 URI.

The s3:// URI can directly be used with the S3 command line tools.
The path-style can not, it is not supported by the "aws s3" cli.