add --require-clean-git-worktree flag to run and status commands
Add a new flag to "baur run" and "baur status" called
--require-clean-git-worktree.
If passed, the commands fail if the baur repository contains untracked or
modified tracked git files, this includes files that are in .gitignore.

"baur run", checks the state on start and before executing a task, to catch
cases were multiple tasks are run and of them modifies or creates new files in
the git repository.

This helps to ensure that tasks are executed in an clean reproducible
environment.