inputfiles: track symlink target paths
baur did not recognize when a input file was a symlink and:
- the target path of a symlink changed to another file with the same content
  then before,
- a symlink was replaced with the target that is was pointing to.

To be able to recognize those changes, the symlink target path must be tracked
as part of the InputFile digest.

The git repository already stores the information if a file is a symlink or not.
This information is utilized when the baur repository is also a git repository
by making git.TrackedObjects available to the InputResolver and looking up the
file type before creating an InputFile.
In a follow-up git.TrackedObjects will also be used to lookup file permissions
when tracking them.

When the baur repository is not a git repository or the file is untracked,
os.Lstat is called to figure out if a path is a symlink and then to lookup the
target path.