gitobject: query db always for resolved symlink paths
If a path for that the digest was requested contained a directory symlink, the
file was not found in the db, despite it contains an entry for the symlink
target path.

gitobject only resolved paths to symlinks if "git ls files" returned a symlink
stauts for them. It's only returned for symlinks though, not for paths to a file
that contain a directory symlink path.

If a path can not be found in the db, always resolve symlink in it's path and
check if an entry for the resolved path exist.